functions-python/requirements.txt
# Avanza Python Backend Dependencies
# Firebase Cloud Functions (Gen 2)
functions-framework==3.*
firebase-functions>=0.1.0
firebase-admin>=6.2.0

# ADK / Gemini
google-adk>=0.10.0
google-genai>=1.3.0

# Market Data
yfinance>=0.2.36
mplfinance>=0.12.10b0
pandas>=2.1.0

# Avanza API Wrapper (unofficial)
avanza-api>=2.0.0

# TOTP Authentication
pyotp>=2.9.0

# HTTP client (for custom requests if needed)
requests>=2.31.0

# Google Cloud (optional - for Firestore access)
google-cloud-firestore>=2.23.0

# YouTube transcripts
youtube-transcript-api>=0.6.2
